Refactor API into route-group modules
