+++ pete_e/api_routes/__init__.py
@@ -1,0 +1,13 @@
+from pete_e.api_routes.logs_webhooks import router as logs_webhooks_router
+from pete_e.api_routes.metrics import router as metrics_router
+from pete_e.api_routes.plan import router as plan_router
+from pete_e.api_routes.root import router as root_router
+from pete_e.api_routes.status_sync import router as status_sync_router
+
+__all__ = [
+    "logs_webhooks_router",
+    "metrics_router",
+    "plan_router",
+    "root_router",
+    "status_sync_router",
+]
